widgetize the join button.  needed to retire Evaluation wiki
